+++ yacg.py
@@ -1,7 +1,7 @@
 import argparse
 import sys
 import logging
-
+from datetime import datetime
 from yacg.util.fileUtils import doesFileExist
 from yacg.util.outputUtils import printError, getErrorTxt, getOkTxt
 from yacg.builder.jsonBuilder import getModelFromJson
@@ -13,6 +13,7 @@ from yacg.model.model import DictionaryType, EnumType, ComplexType
 import yacg.util.yacg_utils as yacg_utils
 import yacg.model.config as config
 import yacg.model.modelFuncs as modelFuncs
+import yacg.util.protocol_funcs as protocolFuncs
 
 
 description = """Yet another code generation.
@@ -39,6 +40,10 @@ parser.add_argument('--vars', nargs='+', help='variables that are passed to the 
 parser.add_argument('--usedFilesOnly', help='import models but only print the used files to stdout', action='store_true')
 parser.add_argument('--flattenInheritance', help='flatten included types so that inheritance', action='store_true')
 parser.add_argument('--noLogs', help='do not print logs', action='store_true')
+parser.add_argument('--protocolFile', help='where the metadata of the used models for this specifig gen job are stored')
+parser.add_argument('--skipCodeGenIfVersionUnchanged', help='when the model versions are unchanged, then the codegen is skipped', action='store_true')  # noqa: E501
+parser.add_argument('--skipCodeGenIfMd5Unchanged', help='when the model file md5 is unchanged, then the codegen is skipped', action='store_true')  # noqa: E501
+parser.add_argument('--skipCodeGenDryRun', help='prints only the log messages if codegen should be skipped', action='store_true')
 
 
 def getFileExt(fileName):
@@ -294,11 +299,40 @@ def _isConfigurationValid(codeGenerationJobs):
 def __doCodeGen(codeGenerationJobs, args):
     """process the jobs to do the actual code generation
     """
-
+    previousCodeGenMetaData = protocolFuncs.getPreviousMetaData(args.protocolFile, args.noLogs)
+    previousJobsMetaData = previousCodeGenMetaData.get("jobs", {})
+    codeGenMetaData = {}
+    jobsMetaData = {}
+    codeGenMetaData["date"] = datetime.now().strftime("%d-%m-%Y %H:%M:%S.%f")
+    codeGenMetaData["jobs"] = jobsMetaData
+    allSkipped = True
+    jobIndex = 1
     for job in codeGenerationJobs:
         alloadedTypes = readModels(job, args.flattenInheritance)
+        modelMetaData = protocolFuncs.getModelMetaData(alloadedTypes, job.models[0].schema)
+        jobName = job.name if job.name else "UNKNOWN_JOB_{}".format(jobIndex)
+        jobsMetaData[jobName] = modelMetaData
+        jobIndex = jobIndex + 1
+        if protocolFuncs.shouldSkipCodeGen(
+                args.skipCodeGenIfVersionUnchanged,
+                args.skipCodeGenIfMd5Unchanged,
+                previousJobsMetaData,
+                modelMetaData,
+                jobName,
+                args.noLogs) is True:
+            if not args.noLogs:
+                logging.info(" SKIP CODEGEN: {}".format(jobName))
+            continue
+        if not args.noLogs:
+            logging.info(" do codeGen: {}".format(jobName))
+        if args.skipCodeGenDryRun is True:
+            if not args.noLogs:
+                logging.info(" 'skipCodeGenDryRun' is set, so no codeGen is executed': {}".format(jobName))
+            continue
+
         # dictionary types are not really useful as toplevel types ... so it's
         # better to remove them - TODO add a commandline switch for that
+        allSkipped = False
         loadedTypes = []
         for t in alloadedTypes:
             if not isinstance(t, DictionaryType):
@@ -322,6 +356,8 @@ def __doCodeGen(codeGenerationJobs, args):
                     task.blackListed,
                     task.whiteListed,
                     task.randomDataTask)
+    if (not allSkipped) and (args.skipCodeGenDryRun is not True):
+        protocolFuncs.writeProtocolFile(args.protocolFile, codeGenMetaData)
 
 
 def __printUsedFiles(codeGenerationJobs, args):