+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 4.3.0
+* yacg commandline switch to protocol the codegen
+* yacg commandline switch to skip codegen when model version hasn't change to the last protocolled run
+
+# 4.2.0
+* add typescript helper funcs
+* enable plantuml template to visualize type and property tags
+* introduce handling of multidimensional arrays
+
+# 4.1.2
+* fixed bug for x-enumValues
+
 # 4.1.1
 * add key related model funcs
 * fix bug with 'additionalProperties'. From JSON schema specs this value can be boolean or object. The boolean version can't used before.